Showing menu only on mouse up
